+++ openair2/LAYER2/MAC/l1_helpers.c
@@ -41,7 +41,7 @@
 #include "UTIL/LOG/log.h"
 #include "proto.h"
 
-int8_t get_Po_NOMINAL_PUSCH(module_id_t module_idP,int CC_id) {
+int8_t get_Po_NOMINAL_PUSCH(module_id_t module_idP,uint8_t CC_id) {
   RACH_ConfigCommon_t *rach_ConfigCommon = NULL;
 
   if (CC_id>0) {
@@ -61,7 +61,7 @@ int8_t get_Po_NOMINAL_PUSCH(module_id_t module_idP,int CC_id) {
 	 get_DELTA_PREAMBLE(module_idP,CC_id));
 }
 
-int8_t get_deltaP_rampup(module_id_t module_idP,int CC_id) {
+int8_t get_deltaP_rampup(module_id_t module_idP,uint8_t CC_id) {
 
   if (CC_id>0) {
     LOG_E(MAC,"Transmission on secondary CCs is not supported yet\n");
